+++ .github/workflows/java.yml
@@ -3,7 +3,7 @@ name: Scala and Java build
 on:
   push:
     branches:
-      - '*'
+      - master
   pull_request:
     branches:
       - '*'
