+++ .github/workflows/python.yml
@@ -3,7 +3,7 @@ name: Python build
 on:
   push:
     branches:
-      - '*'
+      - master
   pull_request:
     branches:
       - '*'
