+++ common/src/main/java/org/apache/sedona/common/Functions.java
@@ -1,14 +1,14 @@
-package org.apache.sedona;
+package org.apache.sedona.common;
 
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Geometry;
 
 public class Functions {
-    public static double ST_Distance(Geometry left, Geometry right) {
+    public static double distance(Geometry left, Geometry right) {
         return left.distance(right);
     }
 
-    public static double ST_YMin(Geometry geometry) {
+    public static double yMin(Geometry geometry) {
         Coordinate[] points = geometry.getCoordinates();
         double min = Double.MAX_VALUE;
         for(int i=0; i < points.length; i++){
@@ -17,7 +17,7 @@ public class Functions {
         return min;
     }
     
-    public static double ST_YMax(Geometry geometry) {
+    public static double yMax(Geometry geometry) {
         Coordinate[] points = geometry.getCoordinates();
         double max = Double.MIN_VALUE;
         for (int i=0; i < points.length; i++) {
